requirements.txt
streamlit
PyMuPDF
openai
Pillow
pytesseract